Refactor for better component naming
