+++ src/app/home/home.module.ts
@@ -6,7 +6,6 @@ import { HttpClientModule, HttpClientJsonpModule } from '@angular/common/http';
 import { HomeComponent } from './home.component';
 import { DashboardComponent } from '../dashboard/dashboard.component';
 import { HomeRouterModule } from './home-router.module';
-import { CommunityComponent } from '../community/community.component';
 import { PlanetFormsModule } from '../shared/forms/planet-forms.module';
 
 import { MaterialModule } from '../shared/material.module';
@@ -16,7 +15,6 @@ import { PlanetDialogsModule } from '../shared/dialogs/planet-dialogs.module';
 import { PulsateIconDirective } from './pulsate-icon.directive';
 import { UpgradeComponent } from '../upgrade/upgrade.component';
 import { SharedComponentsModule } from '../shared/shared-components.module';
-import { CommunityTableComponent } from '../community/community-table.component';
 import { UsersAchievementsModule } from '../users/users-achievements/users-achievements.module';
 import { NewsModule } from '../news/news.module';
 import { LogsMyPlanetComponent } from '../logs-myplanet/logs-myplanet.component';
@@ -41,8 +39,6 @@ import { TeamsModule } from '../teams/teams.module';
   declarations: [
     HomeComponent,
     DashboardComponent,
-    CommunityComponent,
-    CommunityTableComponent,
     DashboardTileComponent,
     NotificationsComponent,
     PulsateIconDirective,
