+++ src/app/manager-dashboard/manager-dashboard.component.html
@@ -6,7 +6,7 @@
 </mat-toolbar>
 <div class="space-container">
   <span *ngIf="planetType !== 'community'">
-    <a routerLink="/requests" i18n mat-raised-button *planetAuthorizedRoles="'manager'">{ planetType, select, nation {Manage Communities} center {Manage Nations} }</a>
+    <a routerLink="requests" i18n mat-raised-button *planetAuthorizedRoles="'manager'">{ planetType, select, nation {Manage Communities} center {Manage Nations} }</a>
     <a [routerLink]="['reports']" i18n mat-raised-button>Reports</a>
   </span>
   <a [routerLink]="['reports', 'myplanet']" i18n mat-raised-button>myPlanet Reports</a>
