+++ src/app/manager-dashboard/requests/requests.component.html
@@ -36,7 +36,7 @@
     </mat-toolbar-row>
   </mat-toolbar>
   <div class="view-container view-full-height" [ngClass]="{'view-table':shownStatus==='pending'}">
-    <planet-community-table *ngIf="shownStatus==='pending'" [data]="filteredData" (requestUpdate)="getCommunityList()"></planet-community-table>
+    <planet-requests-table *ngIf="shownStatus==='pending'" [data]="filteredData" (requestUpdate)="getCommunityList()"></planet-requests-table>
     <ng-container *ngIf="shownStatus==='accepted'">
       <mat-expansion-panel *ngFor="let hub of hubs">
         <mat-expansion-panel-header>
@@ -45,13 +45,13 @@
             <mat-icon>delete</mat-icon>Delete Network
           </a>
         </mat-expansion-panel-header>
-        <planet-community-table [hubs]="hubs" [hub]="hub" [data]="hub.children" (requestUpdate)="getCommunityList()"></planet-community-table>
+        <planet-requests-table [hubs]="hubs" [hub]="hub" [data]="hub.children" (requestUpdate)="getCommunityList()"></planet-requests-table>
       </mat-expansion-panel>
       <mat-expansion-panel>
         <mat-expansion-panel-header>
           <mat-panel-title>Sandbox</mat-panel-title>
         </mat-expansion-panel-header>
-        <planet-community-table [hubs]="hubs" [data]="sandboxPlanets" (requestUpdate)="getCommunityList()"></planet-community-table>
+        <planet-requests-table [hubs]="hubs" [data]="sandboxPlanets" (requestUpdate)="getCommunityList()"></planet-requests-table>
       </mat-expansion-panel>
     </ng-container>
   </div>
