+++ src/app/manager-dashboard/requests/requests.component.spec.ts
@@ -1,10 +1,10 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormsModule } from '@angular/forms';
-import { CommunityComponent } from './community.component';
-import { CouchService } from '../shared/couchdb.service';
+import { CommunityComponent } from './requests.component';
+import { CouchService } from '../../shared/couchdb.service';
 import { HttpClientModule } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { MaterialModule } from '../shared/material.module';
+import { MaterialModule } from '../../shared/material.module';
 import { RouterTestingModule } from '@angular/router/testing';
 import { of } from 'rxjs/observable/of';
